Add PerfMon Button + Cleanup
Add button to reset perfmon counters on SQL nodes.  This will check the
remote node for the .ini file which is used in resetting counters.  Some
other random cleanup: force minimize all other applications on startup,
clear ErrDEPServers on start of routine, and show the main GUI if the
user closes one of the secondary GUI's.