Add robustness to perfmon reset
Check SQL version, ask user if the default paths are used and have them
fill out the path if it's not the default instance name.

Still need to sort out how to parse the instance name out.  Most of the
time we work with the default instance name, but some customers prefer
to name their instances something different.

**Tested on SQL2014 (Win2012R2) and SQL2008R2 (Win2008R2).