Reboot Check
Check to see if you want to reboot the server you are on... I imagine
this may create issues if you have a lot of servers in the list, unless
the server you are on is at the END of the list.  I would imagine the
reboot on the current node would happen before the remaining nodes were
all rebooted.