Service start/stop to Verint
Update the start/stop routine for services to work with Verint instead
of NICE.  I needed to adjust how I was querying services, because in
NICE every service on the node needed to be stopped/disabled or
started/auto.  Since Verint installs all of the services from the
platform, there are services which I should not be touching.  Now the
routine looks for services set to 'auto' and builds a list based on that
- so you should run the 'stop' routine before running the 'start'
routine on any site.